financial-analyzer-python-backend/requirements.txt
boto3==1.35.49
botocore==1.35.49
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
et-xmlfile==1.1.0
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-Cors==5.0.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
openpyxl==3.1.0
pandas==2.2.1
pymongo==3.11.4
python-dateutil==2.8.2
python-dotenv==0.17.1
pytz==2021.1
requests==2.31.0
six==1.16.0
urllib3==1.26.5
Werkzeug==2.0.1
XlsxWriter==1.4.3
yapf==0.31.0
numpy==1.23.0
# uWSGI==2.0.19.1
tensorflow==2.11.0
transformers==4.33.2
nltk==3.8.1
PyJWT==2.7.0
google-cloud-storage==2.14.0
google-api-python-client==2.113.0
google-cloud-aiplatform==1.39.0
gcsfs==2023.12.2.post1
html2text==2020.1.16
chromadb==0.4.12
pdfplumber==0.11.4
pdf2image==1.17.0
pytesseract==0.3.13
PyPDF2==3.0.1
PyMuPDF==1.24.12